[logging] De-stack from the phase-gauge PR
ScalarGauges had no functional dependency on TrainingPhaseGauge; the
stack existed only because both classes shared phase_metrics.py. Move
ScalarGauges to its own module (scalar_gauges.py, one module per
concern) and revert the phase-gauge content from this branch's tree, so
this PR is independently mergeable and its diff is buffer-only. Forward
revert, no history rewrite; #1923 still owns the phase gauge.
